import changeset from labx tree: https://github.com/Biamp-Systems/mb-linux/commit/fb727a9cf5c7f1056b5ab1a562273d5b506f7f31

+++ uClinux-dist/linux-2.6.x/drivers/net/labx_ptp/labx_ptp.h
@@ -732,7 +732,7 @@ struct ptp_device {
   /* RTC control loop persistent values */
   int64_t  integral;
   int64_t  zeroCrossingIntegral;
-  int32_t  derivative;
+  int64_t  derivative;
   int32_t  previousOffset;
   uint32_t prevBaseRtcIncrement;
   uint32_t prevAppliedRtcIncrement;
